% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.data.R
\docType{data}
\name{example.data}
\alias{example.data}
\title{Some example data}
\format{
A list including:
t.data: Tensor data in the target domain with n=50 and (p1,p2,p3)=(10,10,10).
A.data: Tensor data in auxiliary domains with K=5 and nk=100.
t.Omega.true.list: A list, the true precision matrices of all modes in the target domain.
}
\source{
Simulated data
}
\description{
Some example data
}
